finance/market-data/requirements.txt
websocket-client
psycopg2
